+++ jsonFiles/userInput.json
@@ -1,0 +1,25 @@
+[
+
+  {
+    "commands": [
+      "pick up",
+      "go",
+      "punch",
+      "kick",
+      "use",
+      "run",
+      "walk",
+      "restore"
+    ],
+    "objects": [
+      "north",
+      "south",
+      "east",
+      "west",
+      "m4",
+      "neutralizer",
+      "potion"
+    ]
+  }
+
+]
