+++ pom.xml
@@ -74,6 +74,11 @@
       <version>2.8.6</version>
     </dependency>
 
+    <dependency>
+      <groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+      <version>2.8.6</version>
+    </dependency>
 <!--
     To use JUnit4 instead of JUnit5, comment out the previous 2 dependency elements and uncomment
     the following 2 elements.
